fix: move types to prevent dependency cycle
