+++ src/elements/embed/Callout.tsx
@@ -5,7 +5,7 @@ import React, { useEffect, useState } from "react";
 import { Error } from "../../editorial-source-components/Error";
 import { Label } from "../../editorial-source-components/Label";
 import { FieldLayoutVertical } from "../../editorial-source-components/VerticalFieldLayout";
-import type { Campaign } from "../callout/Callout";
+import type { Campaign } from "../callout/CalloutTypes";
 import { EmbedTestId } from "./EmbedForm";
 
 type Props = {
